+++ src/test/testData/oc3/config.php
@@ -1,0 +1,2 @@
+<?php
+define('DB_PREFIX', 'shop_');